moved email_streamer package, timezone package, kafkautil package, forms package and logger package to pkg folder
