+++ pkg/email_streamer/models.go
